+++ extensions/t-ly-url-shortener/CHANGELOG.md
@@ -1,7 +1,7 @@
 # T.LY URL Shortener Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+## [Initial Release] - {PR_MERGE_DATE}
 
-- Create and copy a short link from a URL form.
+- Create and copy a short link from Raycast arguments.
 - Shorten the URL currently on the clipboard.
 - Support an optional default custom short domain.